simple_one_for_one/supervised subscriber_worker implementation is done. It works for normal shutdowns and connection/channel failures.
One remaining outage is that in restarting a channel/connection/queue the subscriber_pool accounting gets out of whack and the queues don't get deleted during a normal restart. This is indicative of a larger problem, namely that the subscriber_pool is losing track of the workers because it's not involved in the workers' lifecycle after initially starting them via the supervisor:start_child/2 function. Worker failures and subsequent restarts are handled entirely by the supervisor at this point. Not having a hook into all lifecycle events compromises the subscriber_pool's ability to logically manage workers for things like cleaning up resources, such as queues, on a normal shutdown.

There are a couple of ways to handle this. 1) Push all of this into the worker itself; 2) Have the worker notify the pool as it's exiting.